pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "product_gallery_normalizer"
version = "0.1.0"
description = "Desktop tool to normalize product images for e-commerce catalogs."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PySide6>=6.7",
    "Pillow>=11.0",
    "pydantic>=2.0",
]

[project.scripts]
product-gallery-normalizer = "product_gallery_normalizer.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/product_gallery_normalizer"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
